Refactored Portfolio folder, Added account folder
