+++ src/components/Account/LoginValidationHook.tsx
@@ -1,7 +1,7 @@
 import axios from 'axios';
 import { useState } from 'react';
 import { useCookies } from 'react-cookie';
-import { url } from '../api/api';
+import { url } from '../../api/api';
 import { toast } from 'react-toastify';
 import {useHistory} from "react-router-dom";
