+++ src/components/Account/ValidationHook.tsx
@@ -1,6 +1,6 @@
 import axios from 'axios';
 import {useState} from 'react';
-import {url} from '../api/api';
+import {url} from '../../api/api';
 import {toast} from 'react-toastify';
 import {useHistory} from "react-router-dom";
