+++ src/components/Portfolio/PortfolioList/PortfolioListTable.tsx
@@ -3,8 +3,8 @@ import React, { useMemo, useState } from "react";
 import { Table, Button } from "react-bootstrap";
 import { useCookies } from "react-cookie";
 import { toast } from "react-toastify";
-import { portfolioUrl } from "../api/api";
-import DeleteConfirmation from "./CreateOrDeletePortfolio/DeleteConfirmation";
+import { portfolioUrl } from "../../../api/api";
+import DeleteConfirmation from "../PortfolioEdit/DeleteConfirmation";
 import { useHistory } from "react-router-dom";
 
 export const defaultArrows:{[key: string]: string} = {
