+++ src/test/Component.test.tsx
@@ -3,9 +3,9 @@ import { render, screen, waitFor } from "@testing-library/react";
 import * as rcookies from 'react-cookie';
 import { BrowserRouter, Route } from "react-router-dom";
 import Adminpage from "../components/Adminpage";
-import validate from "../components/FormValidation";
+import validate from "../components/Account/FormValidation";
 import HonorAwards from "../components/Honors/HonorAward";
-import PortfolioList from "../components/PortfolioList";
+import PortfolioList from "../components/Portfolio/PortfolioList/PortfolioList";
 import ViewPortfolio from "../components/ViewPortfolio/ViewPortfolio";
 import { fullPortfolioTestObj, testAdminFname, testGeneralUser, testPortfolioList } from "./TestUserComponent";
